requirements.txt
python-telegram-bot>=20.0
anthropic>=0.40.0
openai>=1.0.0
requests>=2.31.0
